Fix/#1: SignupRequest 위치 변경 및 AuthController에 signup 로직 위치 변경
