+++ src/main/java/com/seo2yeon/students/domain/auth/dto/SignupRequest.java
@@ -1,9 +1,8 @@
-package com.seo2yeon.students.domain.user.dto;
+package com.seo2yeon.students.domain.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
-import jakarta.validation.constraints.Size;
 import lombok.Getter;
 
 @Getter
