+++ src/main/java/com/seo2yeon/students/domain/user/controller/UserController.java
@@ -1,12 +1,6 @@
 package com.seo2yeon.students.domain.user.controller;
 
-import com.seo2yeon.students.domain.user.dto.SignupRequest;
-import com.seo2yeon.students.domain.user.service.UserService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -14,11 +8,5 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/users")
 @RequiredArgsConstructor
 public class UserController {
-    private final UserService userService;
 
-    @PostMapping("/signup")
-    public ResponseEntity<String> signup(@Valid @RequestBody SignupRequest request) {
-        userService.signup(request);
-        return ResponseEntity.ok("회원가입이 완료되었습니다.");
-    }
 }
